# replacer: evaluate format once and improve perf by ~3x