@miekg 2019-07-14
[ Quoting <notifications@github.com> in "Re: [coredns/coredns] replacer: eva..." ]
>I'll add a comment.  The reason we use store a pointer to the slice and not the slice itself is that it saves an alloc in the value => `interface{}` conversion.

this sounds like a really micro optimizations we shouldn't really care about. I 
doubt this makes a dent in the 3x improvement we're seeing here.

Logging will always kill your performance, doing these micro optimizations won't 
make it that much better. Just use the slice directly.

/Miek

--
Miek Gieben