@miekg 2019-07-14
>**TLDR:** I don't think there is much value in returning an error - since this represents a logic bug.

ok, then remove the comment, no-one will read logs or anything, so if we can't 
make this a failure on startup it doesn't matter.